+++ eig.h
@@ -3,7 +3,7 @@
 
 #include <arpack.h>
 
-void smallest_eigenvalues(double *A, a_int N, a_int nev, double *ret_eigenvalues, double *ret_eigenvectors);
-void all_eigenvalues(double *A, int n, double *ret_eigenvalues);
+void smallest_eigenvalues(float *A, a_int N, a_int nev, float *ret_eigenvalues, float *ret_eigenvectors);
+void all_eigenvalues(float *A, int n, float *ret_eigenvalues);
 
 #endif
