+++ norms.c
@@ -3,17 +3,17 @@
 #include "instrumentation.h"
 
 
-double fast_LUT_exp(double x) {
+float fast_LUT_exp(float x) {
     ENTER_FUNC;
     NUM_ADDS(1);
     NUM_MULS(2);
     uint64_t tmp = (1512775*x+1072632447);
     int index = (int) (tmp>>12) & 0xFF;
     EXIT_FUNC;
-    return ((double )(tmp<<32 ))* ADJUSTMENT_LUT[index];
+    return ((float )(tmp<<32 ))* ADJUSTMENT_LUT[index];
 }
 
-double fast_exp(double x) {
+float fast_exp(float x) {
     ENTER_FUNC;
     NUM_MULS(2);
     NUM_DIVS(1);
@@ -21,18 +21,18 @@ double fast_exp(double x) {
     int x1 = (long long) (6051102*x+1056478197);
     int x2 = (long long) (1056478197-6051102*x);
     EXIT_FUNC;
-    return ((double) x1)/((double) x2);
+    return ((float) x1)/((float) x2);
 }
 
 
 
-double l2_norm(double *u, double *v, int dim) {
+float l2_norm(float *u, float *v, int dim) {
     ENTER_FUNC;
     NUM_ADDS(3*dim);
     NUM_MULS(dim);
     NUM_SQRTS(1);
 
-    double norm0, norm1, norm2, norm3, norm4, norm5, norm6, norm7;
+    float norm0, norm1, norm2, norm3, norm4, norm5, norm6, norm7;
     norm0 = norm1 = norm2 = norm3 = norm4 = norm5 = norm6 = norm7 = 0;
     int i;
     for (i = 0; i < dim - 7; i = i+8) {
@@ -50,19 +50,19 @@ double l2_norm(double *u, double *v, int dim) {
         norm0 += (u[i] - v[i]) * (u[i] - v[i]);
     }
 
-//    double norm = babylonian_squareRoot(norm0+norm1+norm2+norm3+norm4+norm5+norm6+norm7);
-    double norm = sqrt(norm0+norm1+norm2+norm3+norm4+norm5+norm6+norm7);
+//    float norm = babylonian_squareRoot(norm0+norm1+norm2+norm3+norm4+norm5+norm6+norm7);
+    float norm = sqrt(norm0+norm1+norm2+norm3+norm4+norm5+norm6+norm7);
     EXIT_FUNC;
     return norm;
 }
 
 
 
-//double l2_norm_squared(double *u, double *v, int dim) {
+//float l2_norm_squared(float *u, float *v, int dim) {
 //    ENTER_FUNC;
 //    NUM_ADDS(3*dim);
 //    NUM_MULS(dim);
-//    double norm = 0;
+//    float norm = 0;
 //    for (int i = 0; i < dim; i++) {
 //        norm += (u[i] - v[i]) * (u[i] - v[i]);
 //    }
@@ -70,12 +70,12 @@ double l2_norm(double *u, double *v, int dim) {
 //    return norm;
 //}
 
-double l2_norm_squared(double *u, double *v, int dim) {
+float l2_norm_squared(float *u, float *v, int dim) {
     ENTER_FUNC;
     NUM_ADDS(3 * dim);
     NUM_MULS(dim);
 
-    double norm0, norm1, norm2, norm3, norm4, norm5, norm6, norm7;
+    float norm0, norm1, norm2, norm3, norm4, norm5, norm6, norm7;
     norm0 = norm1 = norm2 = norm3 = norm4 = norm5 = norm6 = norm7 = 0;
     int i;
     for (i = 0; i < dim - 7; i = i + 8) {
@@ -92,26 +92,26 @@ double l2_norm_squared(double *u, double *v, int dim) {
     for (; i < dim; i++){
         norm0 += (u[i] - v[i]) * (u[i] - v[i]);
     }
-    double norm = norm0+norm1+norm2+norm3+norm4+norm5+norm6+norm7;
+    float norm = norm0+norm1+norm2+norm3+norm4+norm5+norm6+norm7;
     EXIT_FUNC;
     return norm;
 
 }
 
-double gaussian_similarity(double *u, double *v, int dim) {
+float gaussian_similarity(float *u, float *v, int dim) {
     ENTER_FUNC;
     NUM_EXPS(1);
     NUM_MULS(1);
-    double inner = exp(-0.5 * l2_norm_squared(u, v, dim));
+    float inner = exp(-0.5 * l2_norm_squared(u, v, dim));
     EXIT_FUNC;
     return inner;
 }
 
 
-double fast_gaussian_similarity(double *u, double *v, int dim) {
+float fast_gaussian_similarity(float *u, float *v, int dim) {
     ENTER_FUNC;
     NUM_MULS(1);
-    double inner = EXP(-0.5 * l2_norm_squared(u, v, dim));
+    float inner = EXP(-0.5 * l2_norm_squared(u, v, dim));
 
     EXIT_FUNC;
     return inner;
